+++ lib/src/formatters/escaped_token.dart
@@ -1,0 +1,5 @@
+class EscapedToken {
+  final String text;
+
+  const EscapedToken(this.text);
+}