+++ lib/src/localizations/all.dart
@@ -26,8 +26,8 @@ export 'es_ES.dart';
 import 'fr_FR.dart';
 export 'fr_FR.dart';
 
-import 'zn_CN.dart';
-export 'zn_CN.dart';
+import 'zh_CN.dart';
+export 'zh_CN.dart';
 
 typedef LocalizationFn = MomentLocalization Function();
 
@@ -44,7 +44,7 @@ abstract class MomentLocalizations {
     "es_ES": () => LocalizationEsEs(),
     "fr_FR": () => LocalizationFrFr(),
     "it_IT": () => LocalizationItIt(),
-    "zn_CN": () => LocalizationZnCn(),
+    "zh_CN": () => LocalizationZhCn(),
   };
 
   /// Retreives locale
@@ -165,6 +165,6 @@ abstract class MomentLocalizations {
   /// Italian (Italy)
   static LocalizationItIt it() => LocalizationItIt();
 
-  /// Chinese (China)
-  static LocalizationZnCn znCn() => LocalizationZnCn();
+  /// Simplified Chinese (China)
+  static LocalizationZhCn zhCn() => LocalizationZhCn();
 }